Support permissions using ACL so standard linux permission owner could be someone other than www-data
The main reason for this is to support WSL 2 permissions. The \\wsl$ share does not support ACL permissions
and will automatically set file permissions to the WSL 2 user:usergroup when a file is added or modified using
the \\wsl$ share. So, to still allow Apache to access those files, we are using ACL. Even though www-data isn't
the owner or group on the file, it can still have access through the ACL.